+++ src/test/java/PaymentTest.java
@@ -25,9 +25,12 @@ public class PaymentTest {
 
     @BeforeSuite
     public static void deleteAllFilesFolder() {
-        String path = "/var/lib/jenkins/workspace/Оплата заказа/src/test/resources/";
-        for (File myFile : new File(path).listFiles())
-            if (myFile.isFile()) myFile.delete();
+        objOS_Version = new OS_Version();
+        if (objOS_Version.isUnix()) {
+            String path = "/var/lib/jenkins/workspace/Оплата заказа/src/test/resources/";
+            for (File myFile : new File(path).listFiles())
+                if (myFile.isFile()) myFile.delete();
+        }
     }
 
     @BeforeMethod
@@ -406,10 +409,12 @@ public class PaymentTest {
     @AfterMethod
 
     public void closebrowser(ITestResult testResult) throws IOException {
-        if (testResult.getStatus() == ITestResult.FAILURE) {
-            File scrFile = ((TakesScreenshot)driver).getScreenshotAs(OutputType.FILE);
-            String path = "/var/lib/jenkins/workspace/Оплата заказа/src/test/resources/" + testResult.getName() + ".jpg";
-            FileUtils.copyFile(scrFile, new File(path));
+        if (objOS_Version.isUnix()) {
+            if (testResult.getStatus() == ITestResult.FAILURE) {
+                File scrFile = ((TakesScreenshot) driver).getScreenshotAs(OutputType.FILE);
+                String path = "/var/lib/jenkins/workspace/Оплата заказа/src/test/resources/" + testResult.getName() + ".jpg";
+                FileUtils.copyFile(scrFile, new File(path));
+            }
         }
         driver.quit();
     }
