+++ src/test/java/RegistrationTest.java
@@ -7,6 +7,7 @@ import org.testng.Assert;
 import org.testng.ITestResult;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
 
 import java.io.File;
@@ -26,12 +27,15 @@ public class RegistrationTest {
     RegistrationPage4 objRegistrationPage4;
     static OS_Version objOS_Version;
 
-    /*@BeforeSuite
+    @BeforeSuite
     public static void deleteAllFilesFolder() {
-        String path = "/var/lib/jenkins/workspace/Регистрация покупателя/src/test/resources/";
-        for (File myFile : new File(path).listFiles())
-            if (myFile.isFile()) myFile.delete();
-    }*/
+        objOS_Version = new OS_Version();
+        if (objOS_Version.isUnix()) {
+            String path = "/var/lib/jenkins/workspace/Регистрация покупателя/src/test/resources/";
+            for (File myFile : new File(path).listFiles())
+                if (myFile.isFile()) myFile.delete();
+        }
+    }
 
 
     @BeforeMethod
@@ -337,10 +341,12 @@ public class RegistrationTest {
 
     @AfterMethod
     public void closebrowser(ITestResult testResult) throws IOException {
-        if (testResult.getStatus() == ITestResult.FAILURE) {
-            File scrFile = ((TakesScreenshot)driver).getScreenshotAs(OutputType.FILE);
-            String path = "/var/lib/jenkins/workspace/Регистрация покупателя/src/test/resources/" + testResult.getName() + ".jpg";
-            FileUtils.copyFile(scrFile, new File(path));
+        if (objOS_Version.isUnix()) {
+            if (testResult.getStatus() == ITestResult.FAILURE) {
+                File scrFile = ((TakesScreenshot) driver).getScreenshotAs(OutputType.FILE);
+                String path = "/var/lib/jenkins/workspace/Регистрация покупателя/src/test/resources/" + testResult.getName() + ".jpg";
+                FileUtils.copyFile(scrFile, new File(path));
+            }
         }
         driver.quit();
     }
